requirements.txt
streamlit
opencv-python
torch
torchvision
ultralytics
numpy
Pillow
